No Comments

+++ Assets/Scripts/Teleport.cs
@@ -6,17 +6,19 @@ using UnityEngine;
 
 public class Teleport : MonoBehaviour
 {
+    public bool bluePortal = true;
     [SerializeField] private Transform masterTeleportPoint;
     [SerializeField] private Transform clientTeleportPoint;
    // [SerializeField] private Transform player;
     private void OnTriggerEnter(Collider other)
     {
-        other.gameObject.transform.position = masterTeleportPoint.position;
+        if(other.CompareTag("Carryable"))
+            other.gameObject.transform.position = masterTeleportPoint.position;
         
         if (other.CompareTag("Player"))
         {
             other.gameObject.transform.position = PhotonNetwork.IsMasterClient? masterTeleportPoint.position: clientTeleportPoint.position;
-            if (GameManager.instance != null && PhotonView.Get(other.gameObject).IsMine)
+            if (GameManager.instance != null && PhotonView.Get(other.gameObject).IsMine && bluePortal)
             {
                 GameManager.instance.DeletePuzzle();
                 GameManager.instance.SpawnPuzzleRooms(1);
